Split new text form into separate component, added route so user can use the back button to go back to the texts page.
